requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-RESTX==1.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
Flask-CORS==4.0.0
Flask-Talisman==1.0.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Testing==0.8.1
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
mypy==1.8.0
sentry-sdk==1.39.2

# SEO и оптимизация
Pillow==10.0.0
python-slugify==8.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.16.0
webdriver_manager==4.0.1
pyyaml==6.0.1
webencodings==0.5.1
PyYAML==6.0.1

# Кэширование и Redis
redis==5.0.1
python-memcached==2.0.1

# Аналитика
google-analytics-python==0.1.0
yandex-metrica==1.0.0

# Оптимизация изображений
imageio==2.31.2
imagehash==4.3.1

# Работа с файлами
boto3==1.29.3
pydantic==2.5.3

# Структурированные данные
schema==0.7.5

# Асинхронность
aiohttp==3.9.1
asyncio==3.4.3

# Дополнительные утилиты
python-dateutil==2.8.2
pytz==2023.3

# Сжатие
zstandard==0.21.0
brotlipy==0.7.0

# CDN
cloudflare==5.0.0
fastly==0.1.0

# Геолокация
geopy==2.5.1

# Мониторинг
prometheus-client==0.19.0
statsd==4.0.0
prometheus-flask-exporter==0.21.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-telegram-bot==20.7
PyJWT==2.8.0
redis==5.0.1
aioredis==2.0.1
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
python-multipart==0.0.6
jinja2==3.1.2
alembic==1.12.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-magic==0.4.27
websockets==12.0
openai==1.3.7
langchain==0.0.330
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.12.2
plotly==5.18.0
bokeh==3.2.2
streamlit==1.29.0
jupyter==1.0.0
notebook==7.1.0
